fix(core): widen close_code to u64 and clamp drain semaphore
- CloseInfo.close_code: u32 → u64 to preserve the full QUIC VarInt
  range (up to 2^62 − 1). parse_connection_error no longer truncates.
- session_close: accepts u64 close_code, converts via VarInt::from_u64.
- Drain semaphore: clamp max_concurrency to u32::MAX before cast.
- All adapters (Node, Deno, Tauri) updated for the type change.

Closes #28

+++ packages/iroh-http-deno/src/dispatch.rs
@@ -1083,7 +1083,7 @@ async fn session_next_bidi_stream_dispatch(p: Value) -> Value {
 struct SessionClosePayload {
     endpoint_handle: u32,
     session_handle: u64,
-    close_code: Option<u32>,
+    close_code: Option<u64>,
     reason: Option<String>,
 }
